Handle config block errors
Prior to this commit, having any invalid configuration would cause plug
to halt after encountering an error. If this happened in an early-enough
plugin, it is possible to have no plugins actually load due to this.

This commit implements an info window that tells the user which plugin
the config error originated from, as well as what the actual error was,
when Kakoune opens.